Reorganize primitives code
