+++ Coral/Maze/Maze.js
@@ -1,4 +1,4 @@
-import { GroupPrimitive } from "../../sketchlib/primitives.js";
+import { GroupPrimitive } from "../../sketchlib/rendering/primitives.js";
 import { draw_primitive } from "../../sketchlib/draw_primitive.js";
 import { generate_maze } from "../../sketchlib/RandomDFSMaze.js";
 import { CoralTile } from "../CoralTile.js";
