+++ Coral/Spline.js
@@ -1,4 +1,4 @@
-import { BezierPrimitive } from "../sketchlib/primitives.js";
+import { BezierPrimitive } from "../sketchlib/rendering/primitives.js";
 
 export class Spline {
   constructor(control_points) {
