+++ Worm/AnimatedWorm.js
@@ -5,7 +5,7 @@ import {
   GroupPrimitive,
   LinePrimitive,
   PointPrimitive,
-} from "../sketchlib/primitives.js";
+} from "../sketchlib/rendering/primitives.js";
 import { Style } from "../sketchlib/Style.js";
 import { Color } from "../sketchlib/Color.js";
 import { Motor } from "../pga2d/versors.js";
