+++ lab/ButtonLab/ButtonLab.js
@@ -2,7 +2,10 @@ import { Point } from "../../pga2d/objects.js";
 import { WIDTH, HEIGHT, SCREEN_CENTER } from "../../sketchlib/dimensions.js";
 import { to_y_down } from "../../sketchlib/Direction.js";
 import { draw_primitive } from "../../sketchlib/draw_primitive.js";
-import { CirclePrimitive, GroupPrimitive } from "../../sketchlib/primitives.js";
+import {
+  CirclePrimitive,
+  GroupPrimitive,
+} from "../../sketchlib/rendering/primitives.js";
 import { Style } from "../../sketchlib/Style.js";
 import { CanvasMouseHandler } from "../lablib/CanvasMouseHandler.js";
 import { TouchDPad } from "../lablib/TouchDPad.js";
