+++ lab/PendulumClock/Clock.js
@@ -7,7 +7,7 @@ import {
   GroupPrimitive,
   LinePrimitive,
   VectorPrimitive,
-} from "../../sketchlib/primitives.js";
+} from "../../sketchlib/rendering/primitives.js";
 import { Style } from "../../sketchlib/Style.js";
 import { ClockTime } from "./ClockTime.js";
