+++ lab/lablib/PlayButton.js
@@ -4,7 +4,7 @@ import { HEIGHT, WIDTH } from "../../sketchlib/dimensions.js";
 import {
   GroupPrimitive,
   PolygonPrimitive,
-} from "../../sketchlib/primitives.js";
+} from "../../sketchlib/rendering/primitives.js";
 import { Style } from "../../sketchlib/Style.js";
 import { SCREEN_RECT } from "./Rectangle.js";
 import { TouchButton } from "./TouchButton.js";
