+++ sketchlib/AnimationChain.js
@@ -1,7 +1,11 @@
 import { Point } from "../pga2d/objects.js";
 import { Motor } from "../pga2d/versors.js";
 import { Color } from "./Color.js";
-import { GroupPrimitive, LinePrimitive, PointPrimitive } from "./primitives.js";
+import {
+  GroupPrimitive,
+  LinePrimitive,
+  PointPrimitive,
+} from "./rendering/primitives.js";
 import { Style } from "./Style.js";
 
 const SPINE_STYLE = Style.DEFAULT_STROKE;
